safari-534.53-branch/LayoutTests/fast/js/mozilla/strict/15.3.5.2-expected.txt

FAIL 'use strict'; var f = fn(); delete f.prototype should throw an instance of TypeError
FAIL var f = fn(); delete f.prototype should be false. Was true.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
